Transferring over to React pt1
This is the first half of the React transfer.
As of this commit, the landing page works. You can go to the register page and the base html page is viewable
No slack data is integrated yet, so the base page holds no content.

I'm having an issue where the base page's width is only half that of the screen. Stuck on why right now.

See DummyFile.jsx to learn the React syntax